+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Planned: February 2026
 
 ### Corrected Bugs
 
+* Fixed: [Some devices announce duplicate modules leading to controller error](https://github.com/clicon/clixon-controller/issues/242)
 * Fixed: [device rpc hangs in notification when devices return large amount of data](https://github.com/clicon/clixon-controller/issues/237)
 * Fixed: [Candidate datastore lock prevents sequential commit operations via RESTCONF](https://github.com/clicon/clixon-controller/issues/236)
 * Fixed: [Memory leak after reopening connection](https://github.com/clicon/clixon-controller/issues/169)